requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
django-cors-headers==4.7.0
dj-database-url==2.3.0
gunicorn==20.1.0
psycopg2-binary
whitenoise